// pos/src/app/models/m-pos-master.ts
export class MPosMaster {

    constructor() {}
 

    id?: number;
    posCode?: string;
    posName?: string;
    storeID?: number;
    countryID?: number;
    storegroupID?: number;
    countryName?: string;
    storeGroupName?: string;
    storeName?: string;
    printerdeviceID?; number;
    printerdeviceName?: string;
    defaultCustomer?:string;
    poledisplayPort?: string;
    displaylinemsgOne?: string;
    displaylinemsgTwo?: string;
    diskID?: string;
    cpuID?: string;
    storeCode?: string;
    countryCode?: string;
    storegroupCode?: string;
    defaultcustomerCode?: string;
    active?: boolean;

}